To cran

+++ R/app_label.R
@@ -1,10 +1,56 @@
-#' @title Start app label recordings
-#' @description Starts the app to label recordings
+#' @title Shiny app to label recordings
+#' @description Shiny app to label recordings. Use this app to visualize your
+#' training recordings, create annotations and store them in a sqlite database.
+#' The app has a sidebar panel with the following buttons/boxes to input
+#' required user data: 
+#' \enumerate{
+#'  \item Create database -- if no database exists to store the annotations,
+#'  use this button to create one
+#'  \item Choose database -- choose the database to store the annotations
+#'  \item Butterwoth filter -- indicate low and high frequencies in kHz to 
+#'  filter the recordings
+#'  \item Time expanded -- only used in recorders specifically intended for
+#'  bat recordings. Can take any numeric value. If the recording is not time
+#'  expanded the value must be set to 1. If it is time expanded the numeric
+#'  value corresponding to the time expansion should be indicated
+#'  \item Choose folder -- choose the folder containing the training recordings
+#' }
+#' 
+#' After the spectrogram is ploted:
+#' \enumerate{
+#'  \item Select events by clicking in the spectrogram on the middle of the 
+#'  event of interest (bat call, bird song, etc)
+#'  \item Insert the correct label in the "Label" box and add any additional 
+#'  notes in the "Observations" box
+#'  \item Press 'Set labels' button to add labels to database
+#'  \item Repeat above steps if more than one set of events is present
+#'  in the recording
+#'  \item Press 'Next' button to advance to next recording or pick another 
+#'  recording from the dropdown list
+#'   }
+#' 
+#' The spectrogram can be zoomed by pressing mouse button and dragging to select
+#' an area and then double click on it. To unzoom simply double clicking on the 
+#' spectrogram without an area selected. To adjust visualization settings,
+#' in the top right, the tab "Spectrogram options" can be used to:
+#' \itemize{
+#'  \item Threshold -- minimum intensity values to show
+#'  in the spectrogram. A value of 100 will typically be adequate for the
+#'  majority of the recorders
+#'  \item Window length -- moving window length in ms. Smaller windows best 
+#'  suited for short calls
+#'  \item Overlap -- overlap between consecutive windows, higher values give best 
+#'  visualization but lower performance
+#'  \item Resolution -- frequency resolution of the spectrogram
+#' }
+#' @usage app_label()
+#' @author Bruno Silva
 #' @export
 #' @import htmltools shinyBS
 
 app_label <- function() {
-  ui <- shiny::fluidPage(
+
+    ui <- shiny::fluidPage(
     shiny::titlePanel("Labeler"),
     shiny::sidebarLayout(
       fluid = FALSE,
@@ -79,6 +125,18 @@ app_label <- function() {
             )
           )
         ),
+        shiny::fluidRow(
+          shiny::column(
+          width = 12,
+          align = "center",
+          shiny::numericInput(
+            inputId = "tx",
+            label = "Time expanded",
+            value = "10"
+          )
+          )
+        ),
+
         htmltools::hr(),
         shiny::actionButton(
           inputId = "analysis",
@@ -88,8 +146,8 @@ app_label <- function() {
         htmltools::hr(),
         shinyFiles::shinyDirButton(
           id = "noise_folder",
-          label = "Choose noise folder",
-          title = "Choose noise folder",
+          label = "Add non-relevant",
+          title = "Add non-relevant",
           style = "width:100%"
         )
       ),
@@ -328,15 +386,16 @@ app_label <- function() {
     sound <- shiny::reactive({
       shiny::validate(
         shiny::need(
-          input$files != "",
+         input$files != "",
           "Analysis steps:
-1) Select folder with recordings
-2) If needed, create new database to store recording labels
-3) Select pre-existing database to store recording labels
-4) Select events by clicking in the spectrogram before and after the event of interest (bat call, bird song, etc)
-   Important: Be carefull not to drag mouse while clicking as it will erase previously selected positions
-5) Press 'Set labels' button to add labels to database
-6) Repeat step 4 and 5 if more than one indivudual is in the recording
+1) Select buterrworth filter limits
+2) Input the time expanded factor of the recordings or leave '1' for no time expanded
+3) Select folder with recordings
+4) If needed, create new database to store recording labels
+5) Select pre-existing database to store recording labels
+6) Select events by clicking in the spectrogram on the middle of the event of interest (bat call, bird song, etc)
+7) Press 'Set labels' button to add labels to database
+8) Repeat steps 6 and 7 if more than one set of events is present in the recording
 7) Press 'Next' button to advance to next recording or pick another recording from the dropdown list
 
 Spectrogram visualization:
@@ -345,17 +404,19 @@ Spectrogram visualization:
 - Adjust spectrogram settings with:
     THRESHOLD - minimum energy values displayed, higher values best suited for low quality recordings
     WINDOW - window size in ms, smaller windows best suited for short calls
-    OVERLAP - overlap between consecutive windows, higher values give best visualization but lower performance"
-        )
+    OVERLAP - overlap between consecutive windows, higher values give best visualization but lower performance
+    RESOLUTION - frequency resolution of the spectrogram
+    "
+             )
       )
 
       sound <- import_audio(
         path = input$files,
         butt = TRUE,
         low = as.numeric(input$low),
-        high = as.numeric(input$high)
+        high = as.numeric(input$high),
+        tx = as.numeric(input$tx)
       )
-      print(sound$tx)
       return(sound)
     })
 
@@ -366,18 +427,18 @@ Spectrogram visualization:
       if (!is.null(brush)) {
         ranges$x <- c(brush$xmin, brush$xmax)
         ranges$y <- c(brush$ymin, brush$ymax)
-        file.remove("temp_file.csv")
+        if (file.exists("temp_file.csv")) file.remove("temp_file.csv")
       } else {
         ranges$x <- NULL
         ranges$y <- NULL
-        file.remove("temp_file.csv")
+        if (file.exists("temp_file.csv")) file.remove("temp_file.csv")
       }
     })
     shiny::observeEvent(input$files, {
       ranges$x <- NULL
       ranges$y <- NULL
       maxpos$x <- NULL
-      file.remove("temp_file.csv")
+      if (file.exists("temp_file.csv")) file.remove("temp_file.csv")
     })
 
     shiny::observeEvent(input$specClick$x, {
@@ -390,8 +451,7 @@ Spectrogram visualization:
       )
     })
 
-    output$spec <- shiny::renderPlot(
-      {
+    output$spec <- shiny::renderPlot({
         Spectrogram(
           as.numeric(sound()$sound_samples),
           SamplingFrequency = sound()$fs * sound()$tx,
@@ -456,7 +516,7 @@ Spectrogram visualization:
           tx = shiny::isolate(sound()$tx)
         )
         np <- length(labs)
-        file.remove("temp_file.csv")
+        if (file.exists("temp_file.csv")) file.remove("temp_file.csv")
 
         maxpos$x <- labs
